Get videoframerate and resolution now

+++ BadAppleCMD/Program.cs
@@ -3,7 +3,6 @@ using System.Drawing;
 using System.Globalization;
 using System.Media;
 using System.Text;
-using System.Threading;
 using System.Timers;
 using Timer = System.Timers.Timer;
 
@@ -14,6 +13,7 @@ namespace BadAppleCMD
         private static int _framecounter = 0;
         private static int _totalframecounter = 1;
         private static string _FPS = "FPS: 0";
+        private static Boolean _Desync = false;
         private static SoundPlayer audio;
         private static string strWorkPath = "";
 
@@ -21,6 +21,12 @@ namespace BadAppleCMD
         private static string totalduration;
         private static string currentdurationframe;
         private static Boolean loadingFinished;
+
+        //Video information
+        private static int _VideoWidth;
+        private static int _VideoHeight;
+        private static int _VideoFrameRate;
+
         static void Main(string[] args)
         {
             AppDomain.CurrentDomain.ProcessExit += new EventHandler(ExitApplication);
@@ -30,6 +36,7 @@ namespace BadAppleCMD
 
             //todo add .jpg option? could we do that in a settings page before we start or pass as arg? Maybe both, same for colour
             //todo make sure we calculate window size first and set it (probably not needed)
+            //todo when no file is dropped, allow to input url?
             if (args.Length != 0)
             {
                 path = Path.GetDirectoryName(args[0])
@@ -39,15 +46,10 @@ namespace BadAppleCMD
             }
             else
             {
-                Console.BackgroundColor = ConsoleColor.DarkRed;
-                Console.Clear();
-                string errorstring = "No file has been provided";
-                Console.SetCursorPosition((Console.WindowWidth - errorstring.Length) / 2, Console.WindowHeight / 2 - 3);
-                Console.WriteLine(errorstring);
-
-                string secondlineerrorstring = "Drag and drop a video onto the exe";
-                Console.SetCursorPosition((Console.WindowWidth - secondlineerrorstring.Length) / 2, Console.WindowHeight / 2);
-                Console.WriteLine(secondlineerrorstring);
+                //todo clear this out for release
+                WriteScreen(ConsoleColor.DarkRed, "No file has been provided", "Drag and drop a video on the executable. The program will close in 5 seconds.");
+                //Thread.Sleep(5000);
+                //Environment.Exit(0);
             }
             //Prepare to get every frame
             string strExeFilePath = AppContext.BaseDirectory;
@@ -58,12 +60,28 @@ namespace BadAppleCMD
             di.Attributes = FileAttributes.Directory | FileAttributes.Hidden;
 
             //Save frames and get audio
+            //todo convert to use method
             Console.BackgroundColor = ConsoleColor.DarkBlue;
+            Console.Clear();
+            string MainString = "Now Playing:";
+            Console.SetCursorPosition((Console.WindowWidth - MainString.Length) / 2, Console.WindowHeight / 2 - 3);
+            Console.WriteLine(MainString);
+
+            Console.SetCursorPosition((Console.WindowWidth - path.Length) / 2, Console.WindowHeight / 2);
+            Console.WriteLine(path);
+            Thread.Sleep(2000);
+
+            //Get information about video
+            //todo loading bar but instead with information WOW!
+            string parameter = "-v error -select_streams v:0 -show_entries stream=width,height,avg_frame_rate -of default=nw=1 " + path;
+            Task task = Task.Run(() => { Execute(".\\ffprobe.exe", parameter, true); });
+            LoadingBar("Getting information about the video...");
+
             //TODO Add nifty loading bar instead of outputting it. allow for verbose as arg/option
-            string parameter = "-i " + path + " " + strWorkPath + "/temp/%04d.png";
+            parameter = "-i " + path + " " + strWorkPath + "/temp/%04d.png";
             parameter = parameter.Replace("\\", "/");
             Console.CursorVisible = false;
-            Task task = Task.Run(() => { Execute(".\\ffmpeg.exe", parameter); });
+            task = Task.Run(() => { Execute(".\\ffmpeg.exe", parameter, true); });
             LoadingBar("Getting every frame from the video...");
 
             parameter = "-i " + path + " " + strWorkPath + "/temp/audio.wav";
@@ -72,7 +90,7 @@ namespace BadAppleCMD
             //todo do same as getting video file
             if (!File.Exists(strWorkPath + "/temp/audio.wav"))
             {
-                Execute(".\\ffmpeg.exe", parameter);
+                task = Task.Run(() => { Execute(".\\ffmpeg.exe", parameter, false); });
                 LoadingBar("Getting audio from the video...");
             }
 
@@ -100,15 +118,16 @@ namespace BadAppleCMD
             //todo get fps from video to use as values here
             while (_totalframecounter < fCount)
             {
-                if (_framecounter != 30)
+                if (_framecounter != _VideoFrameRate)
                 {
+                    //todo calculate how many 0s are needed
                     using (Bitmap image = new Bitmap(strWorkPath + $"\\temp\\{_totalframecounter:0000}.png"))
                     {
                         Console.Write(ConvertToAscii(image));
                     }
                     Console.WriteLine(_FPS);
                     File.Delete(strWorkPath + $"\\temp\\{_totalframecounter:0000}.png");
-                    Thread.Sleep(15);
+                    Thread.Sleep(_VideoFrameRate/2);
                     Console.SetCursorPosition(0, 0);
                     _framecounter++;
                     _totalframecounter++;
@@ -117,8 +136,20 @@ namespace BadAppleCMD
             Environment.Exit(0);
         }
 
+        private static void WriteScreen(ConsoleColor BackgroundColour, string MainString, string SecondString)
+        {
+            Console.BackgroundColor = BackgroundColour;
+            Console.Clear();
+            Console.SetCursorPosition((Console.WindowWidth - MainString.Length) / 2, Console.WindowHeight / 2 - 3);
+            Console.WriteLine(MainString);
+            Console.SetCursorPosition((Console.WindowWidth - SecondString.Length) / 2, Console.WindowHeight / 2);
+            Console.WriteLine(SecondString);
+        }
+
         private static void LoadingBar(string MainString)
         {
+            //todo add additional option to display a third line
+            //todo Use WriteScreen() to save code
             Console.Clear();
             while (!loadingFinished)
             {
@@ -144,7 +175,6 @@ namespace BadAppleCMD
                     {
                         //ffmpeg hasnt started properly yet, ignore
                     }
-
                 }
 
                 StringBuilder loadingbar = new("[");
@@ -167,15 +197,15 @@ namespace BadAppleCMD
                     }
                 }
                 loadingbar.Append("]");
-
                 Console.SetCursorPosition((Console.WindowWidth - loadingbar.ToString().Length) / 2, Console.WindowHeight / 2);
                 Console.WriteLine(loadingbar.ToString());
             }
         }
 
-        private static void Execute(string exePath, string parameters)
+        private static void Execute(string exePath, string parameters, Boolean getinformation)
         {
             string result = String.Empty;
+            loadingFinished = false;
 
             using Process p = new();
             p.StartInfo.UseShellExecute = false;
@@ -187,21 +217,49 @@ namespace BadAppleCMD
 
             p.OutputDataReceived += new DataReceivedEventHandler((s, e) =>
             {
-                //Console.WriteLine(e.Data);
+                ///Console.WriteLine(e.Data);
+                if (e.Data is not null)
+                {
+                    if (exePath.Contains("ffprobe.exe") && getinformation)
+                    {
+                        if (e.Data.Contains("width="))
+                        {
+                            _VideoWidth = int.Parse(e.Data.Substring(e.Data.LastIndexOf("width=") + 6));
+                        }
+
+                        if (e.Data.Contains("height="))
+                        {
+                            _VideoHeight = int.Parse(e.Data.Substring(e.Data.LastIndexOf("height=") + 7));
+                        }
+
+                        if (e.Data.Contains("avg_frame_rate="))
+                        {
+                            string frameratefraction = e.Data.Substring(e.Data.LastIndexOf("avg_frame_rate=") + 15);
+                            int valueOne = int.Parse(frameratefraction.Split('/')[0]);
+                            int valueTwo = int.Parse(frameratefraction.Substring(frameratefraction.LastIndexOf('/') + 1));
+                            _VideoFrameRate = valueOne / valueTwo;
+                        }
+                    }
+                }
             });
+
             p.ErrorDataReceived += new DataReceivedEventHandler((s, e) =>
             {
-                //Console.WriteLine(e.Data);
-                if(e.Data is not null)
+                ///Console.WriteLine(e.Data);
+                if (e.Data is not null)
                 {
-                    if (e.Data.Contains("Duration:"))
+                    if (exePath.Contains("ffmpeg.exe"))
                     {
-                        totalduration = e.Data.Substring(e.Data.LastIndexOf("Duration:") + 10, e.Data.LastIndexOf("Duration:") + 9);
+                        if (e.Data.Contains("Duration:") && getinformation)
+                        {
+                            totalduration = e.Data.Substring(e.Data.LastIndexOf("Duration:") + 10, e.Data.LastIndexOf("Duration:") + 9);
+                        }
+                        if (e.Data.Contains("time="))
+                        {
+                            currentdurationframe = e.Data.Substring(e.Data.LastIndexOf("time=") + 5, e.Data.LastIndexOf("time=:") + 12);
+                        }
                     }
-                    if (e.Data.Contains("time="))
-                    {
-                        currentdurationframe = e.Data.Substring(e.Data.LastIndexOf("time=") + 5, e.Data.LastIndexOf("time=:") + 12);
-                    }
+
                 }
             });
 
@@ -209,7 +267,7 @@ namespace BadAppleCMD
             p.BeginOutputReadLine();
             p.BeginErrorReadLine();
             p.WaitForExit();
-            loadingFinished= true;
+            loadingFinished = true;
         }
 
         private static string ConvertToAscii(Bitmap image)
@@ -246,22 +304,38 @@ namespace BadAppleCMD
         private static void OnTimedEvent(Object source, ElapsedEventArgs e)
         {
             _FPS = $"FPS: {_framecounter}";
+            if (_framecounter < _VideoFrameRate && !_Desync)
+            {
+                _Desync = true;
+            }
+            else if (_Desync)
+            {
+                _FPS += " - Video desynced!";
+            }
             _framecounter = 0;
         }
 
         static void ExitApplication(object sender, EventArgs e)
         {
             //todo test every case
-            Console.Clear();
+            //todo different message on program error
+            WriteScreen(ConsoleColor.DarkBlue, "Thank you for playing!", "Cleaning up and exiting...");
+
             //todo check for null
-            audio.Stop();
+            if (audio is not null)
+            {
+                audio.Stop();
+            }
 
             _totalframecounter = 999999999;
             GC.Collect();
             GC.WaitForPendingFinalizers();
 
             //Delete temp folder
-            Directory.Delete(strWorkPath + "/temp", true);
+            if(Directory.Exists(strWorkPath + "/temp"))
+            {
+                Directory.Delete(strWorkPath + "/temp", true);
+            }
         }
     }
 }
